refactor : LocalStorageHabitManager에서 중복해서 저장한 habits 삭제
- LocalStorageHabitManager : 로컬스토리지 사용함을 명시하는 이름으로 변경
- HabitManager 인터페이스와 별도의 파일로 분리
